requirements.txt
fpdf==1.7.2
pdf2image==1.16.2
Pillow==9.4.0
PyPDF2==3.0.1
